Functional components improvements
- defineComponent() now returns function that returns virtual node (with custom type)
- real DOM node creating source code moved to separeted file
- ESLint config updated
